frame/checksum.rs: extract pure 16-bit checksum arithmetic from frame.rs (#988 Phase 1)
First slice of the frame.rs decomposition (per design doc on
research/988-frame-design). Extracts the lowest-coupled cluster:
20 fns of pure 16-bit one's-complement checksum arithmetic for
IPv4/IPv6 + L4 header updates.

Mechanical changes:
- frame.rs → frame/mod.rs (directory module).
- New frame/checksum.rs (~420 LOC): checksum16*, ipv4_words,
  ipv6_words*, adjust_ipv4_header_checksum, checksum16_ipv4/6,
  adjust_l4_checksum_* (12 fns), recompute_l4_checksum_*,
  plus the file-private adjust_l4_checksum_ipv6_addr_bytes
  helper (now pub(super) for the in-file caller in mod.rs).
- frame/mod.rs: declares mod checksum and pub(in crate::afxdp)
  re-exports the cross-module helpers (checksum16, ipv4_words,
  ipv6_words*, adjust_ipv4_header_checksum,
  adjust_l4_checksum_*, recompute_l4_checksum_*).
- Visibility widened pub(super) → pub(in crate::afxdp) on the
  cross-module helpers.

Deferred to later phases per the design doc:
- adjust_l4_checksum_port (line 2312, NAT territory) → frame/nat.rs
- verify_built_frame_checksums (uses build/inspect paths) → frame/build.rs

After this:
- frame/mod.rs: 7,234 LOC (still over the 2,000 threshold; Phase
  2 will extract inspect, then build, then rewrite, then nat).
- frame/checksum.rs: 420 LOC.

cargo test --bins 865/0/2.
